+++ scanners/inspect.py
@@ -2,7 +2,9 @@ import logging
 from scanners import utils
 import json
 import os
-
+import urllib.request
+import base64
+import re
 
 ##
 # == inspect ==
@@ -13,6 +15,42 @@ import os
 
 command = os.environ.get("SITE_INSPECTOR_PATH", "site-inspector")
 init = None
+
+chrome_preload_list = None
+
+
+def get_chrome_preload_list():
+    logging.debug("Fetching Chrome preload list from source...")
+
+    preload_list_url = 'https://chromium.googlesource.com/chromium/src/net/+/master/http/transport_security_state_static.json'
+    preload_list_url_as_text = preload_list_url + '?format=text'
+    with urllib.request.urlopen(preload_list_url_as_text) as response:
+        raw = response.read()
+
+    # To avoid parsing the contents of the file out of the source tree viewer's
+    # HTML, we download it as a raw file. googlesource.com Base64-encodes the
+    # file to avoid potential content injection issues, so we need to decode it
+    # before using it. https://code.google.com/p/gitiles/issues/detail?id=7
+    raw = base64.b64decode(raw).decode('utf-8')
+
+    # The .json file contains '//' comments, which are not actually valid JSON,
+    # and confuse Python's JSON decoder. Begone, foul comments!
+    raw = ''.join([ re.sub(r'//.*$', '', line)
+                    for line in raw.splitlines() ])
+
+    preload_list_json = json.loads(raw)
+    return { entry['name'] for entry in preload_list_json['entries'] }
+
+
+def init(options):
+    """
+    Download the Chrome preload list at the beginning of the scan, and
+    re-use it for each scan. It is unnecessary to re-download the list for each
+    scan because it changes infrequently.
+    """
+    global chrome_preload_list
+    chrome_preload_list = get_chrome_preload_list()
+    return True
 
 def scan(domain, options):
     logging.debug("[%s][inspect]" % domain)
@@ -55,6 +93,7 @@ def scan(domain, options):
         max_age,
         data['hsts_entire_domain'],
         data['hsts_entire_domain_preload'],
+        domain in chrome_preload_list,
         data['broken_root'], data['broken_www']
     ]
 
@@ -65,6 +104,7 @@ headers = [
     "Downgrades HTTPS", "Strictly Forces HTTPS",
     "HTTPS Bad Chain", "HTTPS Bad Hostname",
     "HSTS", "HSTS Header", "HSTS Max Age",
-    "HSTS All Subdomains", "HSTS Preload Ready",
+    "HSTS All Subdomains",
+    "HSTS Preload Ready", "HSTS Preloaded",
     "Broken Root", "Broken WWW"
 ]
